Extract client from provider and filter by currency
