+++ jobs/Backend/Task/src/ExchangeRateProvider/Currency.cs
@@ -1,4 +1,4 @@
-namespace ExchangeRateUpdater;
+namespace ExchangeRateProvider;
 
 public class Currency
 {
