+++ jobs/Backend/Task/src/ExchangeRateProvider/ExchangeRate.cs
@@ -1,4 +1,4 @@
-namespace ExchangeRateUpdater;
+namespace ExchangeRateProvider;
 
 public class ExchangeRate
 {
