+++ jobs/Backend/Task/src/ExchangeRateProvider/Http/IExchangeRateClient.cs
@@ -1,0 +1,7 @@
+﻿using System.Collections.Generic;
+namespace ExchangeRateProvider.Http;
+
+public interface IExchangeRateClient
+{
+    Task<IEnumerable<CurrencyRate>> GetDailyExchangeRates();
+}
