+++ jobs/Backend/Task/src/ExchangeRateProvider/IExchangeRateProvider.cs
@@ -1,10 +1,6 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
+﻿namespace ExchangeRateProvider;
 
-namespace ExchangeRateUpdater
+public interface IExchangeRateProvider
 {
-	public interface IExchangeRateProvider
-	{
-		Task<IEnumerable<ExchangeRate>> GetExchangeRates(IEnumerable<Currency> currencies);
-	}
+	Task<IEnumerable<ExchangeRate>> GetExchangeRates(IEnumerable<Currency> currencies);
 }
