+++ jobs/Backend/Task/src/TestApp/Program.cs
@@ -1,21 +1,8 @@
-using ExchangeRateUpdater;
+using ExchangeRateProvider.Http;
+using ExchangeRateProvider;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-
-
-IEnumerable<Currency> currencies =
-[
-	new Currency("USD"),
-    new Currency("EUR"),
-    new Currency("CZK"),
-    new Currency("JPY"),
-    new Currency("KES"),
-    new Currency("RUB"),
-    new Currency("THB"),
-    new Currency("TRY"),
-    new Currency("XYZ")
-];
 
 
 var configBuilder = new ConfigurationBuilder();
@@ -26,16 +13,31 @@ configBuilder.SetBasePath(Directory.GetCurrentDirectory())
 var builder = Host.CreateDefaultBuilder(args)
     .ConfigureServices((context, services) =>
     {
-        services.AddHttpClient<IExchangeRateProvider, ExchangeRateProvider>(client =>
-        {
-            var baseUrl = context.Configuration["BaseUrl"] ?? throw new Exception("BaseUrl configuration not found.");
+		services
+            .AddSingleton<IExchangeRateProvider, ExchangeRateProvider.ExchangeRateProvider>()
+		    .AddHttpClient<IExchangeRateClient, CnbExchangeRateClient>(client =>
+            {
+                var baseUrl = context.Configuration["BaseUrl"] ?? throw new Exception("BaseUrl configuration not found.");
 
-			client.BaseAddress = new Uri(baseUrl);
-        });
-    });
+                client.BaseAddress = new Uri(baseUrl);
+            });
+	});
 
 var host = builder.Build();
 
+
+IEnumerable<Currency> currencies =
+[
+	new Currency("USD"),
+	new Currency("EUR"),
+	new Currency("CZK"),
+	new Currency("JPY"),
+	new Currency("KES"),
+	new Currency("RUB"),
+	new Currency("THB"),
+	new Currency("TRY"),
+	new Currency("XYZ")
+];
 
 try
 {
